멀티쓰레드 렌더링 버그 수정
. 섀도우 맵의 초기화 시점이 이르기 때문에 그림자가 제대로 렌더링되지 않던 버그 수정
. 섀도우 맵의 초기화를 프레임워크 생성시에서 씬 생성시로 늦추고, 렌더링 패스에서 한번만 섀도우 맵의 초기화를 실시하기 위해 프레임워크에서 씬의 섀도우 맵을 읽어와서 관리
. 그림자가 존재하지 않는 씬에서는 그림자를 렌더링하는 커맨드 리스트가 작동하지 않도록 예외 처리.

+++ 소스/WorldEnd/Client/Client.cpp
@@ -99,7 +99,7 @@ BOOL InitInstance(HINSTANCE hInstance, int nCmdShow)
 {
     hInst = hInstance; // 인스턴스 핸들을 전역 변수에 저장합니다.
 
-    RECT rect{ 0, 0, g_GameFramework.GetWindowWidth(), g_GameFramework.GetWindowHeight() };
+    RECT rect{ 0, 0, (LONG)g_GameFramework.GetWindowWidth(), (LONG)g_GameFramework.GetWindowHeight() };
     DWORD dwStyle{ WS_OVERLAPPED | WS_CAPTION | WS_MINIMIZEBOX | WS_SYSMENU | WS_BORDER };
     AdjustWindowRect(&rect, dwStyle, false);
